Implementing a major fix where the Slots relation wasn't being handled properly and so saves weren't cascading down correctly onto the Slots objects.
There is a new, fairly extensive test to cover this. Overall, this is a huge improvement as the slots don't save to the database until you actually call ->save() on your record. The result is that the slots look and feel exactly like fields. In other words, if you set a field and don't save, you expect that field to not be persisted. The same behavior is true of slots.